R_markdown import lines fixed
The R_markdown code to import temp and demand data was referenced incorrectly. It was changed to work directly with the file format used in the GitHub repository and also to unzip any folders required by itself.
I did not manage to make it work for unzipping the forecast demand for NSW however. This is left for later.