Fixed issues in data cleaning and exported from R to a csv
Data cleaning Rmd needed a lot of changes. There were issues with converting DATETIME the way it was before. using Regex seems to have overcome any issues.
The first round of cleaning is done and the results exported to a csv so that others can start to do some exploratory visualization and basic modelling hopefully.