Reduce the complexity of the forecast_nsw data and store it in the repository. Begin forecast data cleaning.
The original forecast_nsw data was downloaded in 2 parts and corrupt. A workaround was made to download the correct file. I could then remove parts of the data i.e. PREDISPATCHSEQNO and REGIONID that were superfluous on my local machine before zipping the file and saving, this meets the storage requirements of GitHub and allowed storage of the file in the repository, able to be used in the future by the team (this was done in Forecast_File_Reduction.Rmd).
Also began cleaning this forecast data in Forecast_Cleaning.Rmd to potentially be used as an evaluation metric for future model performance.