Started data exploration of NSW data
Looked at the relationship between temperature and demand data for the NSW dataset in order to find any interesting trends. Some interesting concepts were covered in an attempt to fit a simple linear model. The model did not perform well and the next steps will be to find interesting features that may be useful in using to train a ML model.