Add files via upload
Created the Neural Network for Energy Forecast. The Input file is saved in the folder as well (Demand_Temperature_Data.csv). On Exploration of the Consolidated Data I found that Temperature has high correlation with the Energy Data. So thought to use only Temperature as the input parameter to forecast Demand. So only Demand and Temperature data has been used. This Neural Network is trained for 10 Epochs with a Batch size of 1000 and you will see that the Loss is reducing and thus it is converging. Let me know what you guys think.